requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-ai-documentintelligence==1.0.0b4
azure-ai-inference==1.0.0b7
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachelib==0.13.0
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cryptography==42.0.8
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.16.1
Flask==3.0.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-community
langchain-core
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langcodes==3.4.0
langsmith==0.3.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
more-itertools==10.5.0
msal==1.29.0
msal-extensions==1.2.0
msgspec==0.19.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
openai==1.60.0
openpyxl==3.1.5
orjson==3.10.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.0
preshed==3.0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.8.0
pymssql==2.3.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rq==2.1.0
safetensors==0.4.5
shellingham==1.5.4
simple-salesforce==1.12.6
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.37.2
tabulate==0.9.0
tenacity==8.4.2
thinc==8.2.5
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.44.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
zeep==4.2.1
zstandard==0.23.0
